LayoutTests/platform/chromium/inspector/debugger/dom-breakpoints-expected.txt

Tests DOM breakpoints. Bug 42886

Debugger was enabled.

Running: testInsertChild
Test that 'Subtree Modified' breakpoint is hit when appending a child.
Set 'Subtree Modified' DOM breakpoint on rootElement.
Append childElement to rootElement.
Script execution paused.
Call stack:
    0) appendElement (dom-breakpoints.html:12)
    1)  (:1)
Paused on a "Subtree Modified" breakpoint set on div#rootElement, because a new child was added to that node.
Script execution resumed.

Running: testInsertGrandchild
Test that 'Subtree Modified' breakpoint is hit when appending a grandchild.
Append grandchildElement to childElement.
Script execution paused.
Call stack:
    0) appendElement (dom-breakpoints.html:12)
    1)  (:1)
Paused on a "Subtree Modified" breakpoint set on div#rootElement, because a new child was added to its descendant div#childElement.
Script execution resumed.

Running: testRemoveChild
Test that 'Subtree Modified' breakpoint is hit when removing a child.
Remove grandchildElement.
Script execution paused.
Call stack:
    0) removeElement (dom-breakpoints.html:44)
    1)  (:1)
Paused on a "Subtree Modified" breakpoint set on div#rootElement, because its descendant div#grandchildElement was removed.
Script execution resumed.

Running: testModifyAttribute
Test that 'Attribute Modified' breakpoint is hit when modifying attribute.
Set 'Attribute Modified' DOM breakpoint on rootElement.
Modify rootElement data-test attribute.
Script execution paused.
Call stack:
    0) modifyAttribute (dom-breakpoints.html:18)
    1)  (:1)
Paused on a "Attribute Modified" breakpoint set on div#rootElement.
Script execution resumed.

Running: testModifyAttrNode
Test that 'Attribute Modified' breakpoint is hit when modifying Attr node.
Set 'Attribute Modified' DOM breakpoint on rootElement.
Modify rootElement data-test attribute.
Script execution paused.
Call stack:
    0) modifyAttrNode (dom-breakpoints.html:24)
    1)  (:1)
Paused on a "Attribute Modified" breakpoint set on div#rootElement.
Script execution resumed.

Running: testSetAttrNode
Test that 'Attribute Modified' breakpoint is hit when adding a new Attr node.
Set 'Attribute Modified' DOM breakpoint on rootElement.
Modify rootElement data-foo attribute.
Script execution paused.
Call stack:
    0) setAttrNode (dom-breakpoints.html:32)
    1)  (:1)
Paused on a "Attribute Modified" breakpoint set on div#rootElement.
Script execution resumed.

Running: testModifyStyleAttribute
Test that 'Attribute Modified' breakpoint is hit when modifying style attribute.
Set 'Attribute Modified' DOM breakpoint on rootElement.
Modify rootElement style.color attribute.
Script execution paused.
Call stack:
    0) modifyStyleAttribute (dom-breakpoints.html:38)
    1)  (:1)
Paused on a "Attribute Modified" breakpoint set on div#rootElement.
Script execution resumed.

Running: testRemoveNode
Test that 'Node Removed' breakpoint is hit when removing a node.
Set 'Node Removed' DOM breakpoint on elementToRemove.
Remove elementToRemove.
Script execution paused.
Call stack:
    0) removeElement (dom-breakpoints.html:44)
    1)  (:1)
Paused on a "Node Removed" breakpoint set on div#elementToRemove.
Script execution resumed.

Running: testReload
Test that DOM breakpoints are persisted between page reloads.
Set 'Subtree Modified' DOM breakpoint on rootElement.
Page reloaded.
Append childElement to rootElement.
Script execution paused.
Call stack:
    0) appendElement (dom-breakpoints.html:12)
    1)  (:1)
Paused on a "Subtree Modified" breakpoint set on div#rootElement, because a new child was added to that node.
Script execution resumed.
Debugger was disabled.
